+++ kitty/state.h
@@ -414,6 +414,7 @@ typedef struct BorderRect {
     uint32_t color;
     long long border_type;
     bool horizontal;
+    unsigned radius, thickness;
 } BorderRect;
 
 typedef struct BorderRects {
@@ -707,6 +708,7 @@ OSWindow *current_os_window(void);
 void os_window_regions(const OSWindow *, Region *main, Region *tab_bar);
 bool drag_scroll(Window *, OSWindow *);
 void draw_borders(ssize_t vao_idx, unsigned int num_border_rects, BorderRect *rect_buf, bool rect_data_is_dirty, color_type, unsigned int, bool, OSWindow *w);
+void draw_rounded_borders(BorderRects *, color_type, unsigned int, bool, OSWindow *);
 ssize_t create_cell_vao(void);
 ssize_t create_border_vao(void);
 void bind_shader_globals_to_current_context(void);